MySQL connected, Get, Post, Put, Delete

+++ src/main/java/com/example/app/error/exception/TrainNotFoundException.java
@@ -1,7 +1,7 @@
 package com.example.app.error.exception;
 
 public class TrainNotFoundException extends RuntimeException{
-    public TrainNotFoundException(Long id) {
+    public TrainNotFoundException(int id) {
         super("Could not find train " + id);
     }
 }
